@@ -12,18 +12,18 @@
     "@tailwindcss/postcss": "^4.0.17",
     "framer-motion": "^12.6.2",
     "lucide-react": "^0.485.0",
-    "next": "^13.2.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "15.2.4",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@types/node": "^20.17.28",
-    "@types/react": "^18.0.12",
-    "@types/react-dom": "^18",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19",
     "autoprefixer": "^10.4.21",
     "eslint": "^9",
-    "eslint-config-next": "^13.2.4",
+    "eslint-config-next": "15.2.4",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.0.17",
     "typescript": "^5.8.2"
